fix: FORGE-4700 supress alerts for DPUs in certain states
Adds a new classification so that metrics will ignore alerts with the new classification.  This leaves the alert for all other views.

I started looking at removing the classifications for alerts in some states, but they are used for decision making in the state machine so that won't work

See merge request nvmetal/carbide!3299