fix: check serial number in the expected machines manifest against the SKU that is reported from redfish
Dells use the asset tag to identify their machines. The asset tag is a unique identifier which differs from the serial number. As a result, operators will likely specify the asset tag as the "Serial Number" in the expected machines manifest for Dells. This asset tag is identified as the SKU in the System information in the report that site explorer collects for every BMC endpoint.

See merge request nvmetal/carbide!3360