Machine lifecycle test: Update state machine check
Updating the machine lifecycle test to account for changes in the state machine.

Previously we waited 60 mins for the machine to reach HostInitializing/WaitingForDiscovery. However, that is no longer the first HostInit state. Now we will wait for the machine to reach any state whose name starts with 'HostInitializing'.

See merge request nvmetal/carbide!3372