Machine lifecycle test: Improve wait for redfish endpoint
We had a test failure during factory-reset because after a reboot we wait for any successful response from the redfish endpoint before proceeding. However, forge-admin-cli requires a successful response _with data_. There must be a delay between these two, causing the failure.

Improving the wait to include a check for the 'Vendor' field in the response (this is the field forge-admin-cli checks for before submitting a request).

See merge request nvmetal/carbide!3397