feat: Allow to view and set UFM IP in carbide config file
With this change, the UFM IP is no longer hidden in Vault. Instead of that, the config file shows the IP, and lets the site-admin change the IP.

This makes it more obvious what the UFM instance is we are talking to. It also makes it more obvious on how to change the IP (just edit the config file and restart carbide).

The change is based on https://gitlab-master.nvidia.com/nvmetal/carbide/-/merge_requests/2046 from Klaus Ma, and also paves the way for multi-fabric InfiniBand support. However the config in this change only allows to configure a single fabric with the name "default".

Rollout out this change requires some migration steps **if IB support is enabled for the site:**

1. In `carbide-api-site-config.toml`, add an entry for the `default` infiniband fabric.
It should look like
```
[ib_fabrics.default]
endpoints = ["https://1.2.3.4"]
pkeys = [{ start = "1", end = "10" }]
```

The UFM endpoint address for each site can be reconstructed from site vault on each site.
Execute `vault kv get -field=UsernamePassword --tls-skip-verify secrets/ufm/default/auth`
to find the IP.

The pkey range can be taken from the `[pools.pkey]` value in the existing site-config.

2. Remove the top-level `[pools.pkey]` entry.

Since the pkey range is now part of the fabric configuration, it will no longer
be required here. This step is required. If the raw pkey range is not removed, carbide will crash on startup in order to allow us detecting the misconfiguration.

**Note:** With this change, the UFM IP will still be written into Vault when `forge-admin-cli credential add-ufm` is called. However the value written there is ignored.

Implements https://jirasw.nvidia.com/browse/FORGE-4739

See merge request nvmetal/carbide!3394