feat(fnn): add unique asn per dpu
A hacked version of this was running in demo2 for awhile now.  this formalizes the change and keeps compatibility with the ETV VPC.

changes:

1. adds an "optional" ASN pool for FNN.
2. assigns an ASN to each DPU if the pool has free entries
3. uses the assigned ASN when FNN is enabled for a DPU
4. uses the site-wide ASN when FNN is disabled

notes:

* The pool needs to be a u32 in order to fit the range GNI has assigned
  * the resource pool code was using i32 and was updated to i64
  * the database only uses signed numbers
* The dpu specific ASN is assigned for the life of the DPU regardless of the use of FNN
* Upgrade attempts to assign an ASN to a DPU, but leaves the value as NULL if it can't, so that it can be fixed with a config file update
* if you saw a previous version of this, note that there are no longer any changes to VNI as that now comes from the segment.

testing:

manually tested in dev env for backwards compatiblity.  will need to be checked in dev3 or dev4.

manually tested in demo2 for FNN

any unit test suggestions welcome.

See merge request nvmetal/carbide!3407