feat: FORGE-5121: prevent misleading logging
When we are in the `HostInit::Measuring` state, if the `EkCertVerificationStatus` does not exist, this means the `discover_machine()` has not been called yet. We identify that condition and make the state machine wait for it to be called instead of simply complaining that `EkCertVerificationStatus` does not exist as it is misleading.

See merge request nvmetal/carbide!3362