Use JSONB_AGG to fix N+1 queries in various places
There's a bunch of places where in order to get a single object from the
database, we need to query a bunch of related tables. When querying
multiple/all objects we've been avoiding the worst performance pitfalls
by querying related data by list-of-ids and associating it together
after the fact, but there's still issues with things like
network_segments where we have actual N+1 query bugs.

Postgres has a solution here: The JSONB_AGG function. It can bring in
related data, serialized as JSON, in a single query. Sqlx gives decent
integration with sesrde_json, so reading the aggregated JSON back into a
vec of related objects is trivial.

This change creates a set of views for machines, machine_interfaces, and
network_segments, named with a "_snapshots" naming convention, which use
JSONB_AGG to bring in the related data in a single call. For optional
cases like machine_state_history and network_segment_state_history,
this creates a second view "_snapshots_with_history" which joins in the
history as well.

This simplifies a ton of the code so that we don't need a bunch of
`load_related_data` functions all over, and makes it so every code path
gets the related data, not just where we remembered to do it.

The only N+1 issue we still have is in the `num_free_ips` property of
NetworkSegment, where we have to do our own logic to calculate that
number and can't simply put it in a postgres view. (Although we could,
if we wanted to implement the IP allocator as a postgres function...)

See merge request nvmetal/carbide!3435