feat(admin-cli): Add VPC prefix support
This adds a new subcommand `vpc-prefix` to `forge-admin-cli`, with three actions (`create`, `show`, `delete`).

Also included is a change to the API's internal RBAC rules to allow VPC prefix write operations from the CLI (at the moment only the site-agent is allowed to touch these, which is inconvenient for development).

See merge request nvmetal/carbide!3463