fix: Restart KEA when a MachineInterface gets deleted
As reported in https://nvbugspro.nvidia.com/bug/4792034, KEA maintains an internal cache of DHCP records. That cache does not get updated when a
DHCP record is deleted on Carbide.

In order to enforce deletion of the KEA cache, this change will restart KEA whenever a MachineInterface (underlay or admin IP) on Carbide gets deleted.

The change is implemented by
- A new table `machine_interfaces_deletion` records whenever a MachineInterface is deleted
- The deletion time is reported as `last_invalidation_time` whenever a DHCP response is returned by the `DiscoverDhcp` API.
- When `last_invalidation_time` is newer than the startup time of KEA, KEA will restart.

Note that the last_invalidation_time is **not** updated when an Instance DHCP record is updated. As long as instance DHCP records are handled purely by DPUs, this should not be required. In system without DPUs, we might however also need to address this scenario.

Fixes https://nvbugspro.nvidia.com/bug/4792034

See merge request nvmetal/carbide!3476