FORGE-5085: Prevent null org names in db
We had a bug where org names were null in databases from before we
introduced organization_name as a mandatory column in the code, causing
errors on FindTenantsByOrganizationIds. It's not super clear where else
we're going to find this issue, but just in case, populate the data with
a default value if it's null, and make it non-null.

See merge request nvmetal/carbide!3478