feat: Allow to store and fetch Metadata for Machines
With this change, Machines store the same set of metadata that is available for VPCs and instances:
- Name
- Description
- Labels

With this change, a set of default metadata is inserted when the Machine is created. The name is set to the Machine ID of the Machine.
All other fields will be empty.
For existing Machines the name field will be populated in the same
manner.

Follow-up changes will provide the ability to update metadata, as well as to automatically setup metadata during initial site creation.

See merge request nvmetal/carbide!3506