feat: The ExpectedMachines definitions can add Metadata to newly ingested Machines
With this change it is possible to store Metadata within the Expected Machines table and automatically apply it to newly created Machines on ingestion.

It e.g. allows to store asset tags as part of the definition.

In order to implement the functionality, the Metata fields are now transferred as part of the ExpectedMachine RPC definition and admin-cli allows to display and edit the definitions.

See merge request nvmetal/carbide!3509