feat(zero-dpu): Add instance_network_restriction to rpc::Machine
This is to assist cloud-api in allocating instances on zero-dpu hosts in
mixed environments (ones where there are hosts with and without DPUs.)

If a user requests an instance in a VPC that is known to contain
"externally managed" networks (how we tag this in cloud-api is still
TBD), only machines that are "statically" in that VPC should be eligible
to have the instance allocated on them.

Likewise when doing a normal allocation request for an instance with
instances on tenant overlay, cloud-api should *exclude* any hosts with
native_vpc_id set.

To make this work, introduce a new field on rpc::Machine of type
InstanceNetworkRestrictions, which describes:

- Can this machine's interfaces be configured by the tenant, or are they
  static
- What VPC ID the instance's networking should be limited to, if any
- What network segments the instance's networking is limited to, if
  static

See merge request nvmetal/carbide!3413