fix: mkosi use /tmp folder for output instead of git tree
Right now mkosi uses git source tree to save output of the build phase, it creates number of challenges as mkosi uses rootless container which maps `subuid` and `subgid` to the user which does not belongs to the runner itself.

There is possible fixes to that, one involves running `lxc-usernsexec` or `rootlesskit` to fix folder in `afterscript`  another option is to write this output folder outside of the git tree.

This MR takes second approach.

Signed-off-by: ianisimov ianisimov@nvidia.com

See merge request nvmetal/carbide!3493