Internalize static configs for dpus
The dpu firmware configs have caused a lot of problems in our
deployments. They are very complicated and you need to understand a
lot about the internals of forge and what version is deployed in order
to fill the details out properly. These values are always static and
there is only one version for any given deployment. There is no reason
for us to externalize these values in an environment, so this commit
hides all of the messy details from the deployment.

See merge request nvmetal/carbide!3495