feat: have site explorer only update the Admin BMC account's password and keep the factory username
Remove the concept of a site-wide BMC username; only maintain a sitewide BMC password in vault. maintain the username from the expected machines table as the username of the administrator BMC account.

Vendors enforce differing rules on the usernames for admin accounts. For example, Vikings do not allow a username of "root" for an account with administrative privileges

See merge request nvmetal/carbide!3542