Additional Maintenance Mode targets
This change adds health overrride templates that allow operators to add new "Maintenance" health alerts with different targets. The targets are defines as follows:

- `null` (no target): Describes that the host is in maintenance by Forge internal workflows. This mode is used by setting the currently existing Maintenance mode on hosts, as well as by update workflows.
- `"OutForRepair"`: Describes that a Machine is out for repair and requires intervention by an external party.
- `"Degraded"`: Describes that a Machine is still in use by a Tenant, but is known to have issues.

The usage of different targets in Maintenance will allow for better accounting which party is currently required to resolve an issue with a Host.

All Maintenance modes will prevent other tenants from using the Machine as instances - in case the current instance will get freed again.

The new "Degraded" and "OutForRepair" targets are set in the Health Report with source `manual-maintenance` by default, whereas forge internal health reoprts use source `maintenance`. Thereby the internal reports and alerts can always coexist with the Degraded/OutForRepair alerts. This is done for implementation reasons and might be revisited in thee future (internal maintenance mode still being used by AutoUpdater). The main practical implementation is that in case operators don't want the Maintenance mode to be set with 2 different targets, the might need to manually clean the 2nd Maintenance mode target first.

In addition to the new targets, the change will allow for better control which hosts are included in the Prometheus/AlertManager/PagerDuty alerts for unhealthy Machines and which are excluded.

Up to now we used the presence of any `Maintenace` alert for excluding the Machines - but with the alert being used for multiple purposes this might not always be desirable.

With the change, an additional classification with name `SuppressExternalAlerting` will get added to Maintenance alerts. And in the future the presence of the classification can be used for excluding machines from paging thresholds.

With this change, every Machine in Maintenance mode will automatically also get the SuppessPaging classification. However we might want to fine-tune this in the future, and e.g. keep alerting for hosts that are marked in Maintenance for Forge-internal reasons (target: None).

See merge request nvmetal/carbide!3578