Make db::machine::Machine's db methods freestanding functions
This is step 1 of a 2-step process to combining Machine and
MachineSnapshot. This leaves them both in place, but does the following:

- Move the definition of Machine to crate::model::machine
- Take all the database concerns and makes them freestanding functions
  in db::machine
- Remove all the trivial "getter" methods on machine, making the fields
  public, and inlining them at the call site. [0]
- Moves the remaining non-database functions to model::machine::Machine

The next step will be to combine Machine and MachineSnapshot, since
they're basically redundant at this point.

[0] Getters are IMO an anti-pattern in Rust, because callers may want to
clone the data, they may want a reference, or they may want to take
ownership of the field. When making a getter you have to pick one of
these, and it's not always approprate for all callers.

See merge request nvmetal/carbide!3576