Delete dead code from api crate
Symbols which nested deeply in the crate and are `pub`, but are not
exported as pub through `lib.rs`, can sometimes evade rust's dead code
detection, which tends to be conservative.

One way to fix this is to make everything `pub(crate)` instead of pub,
but (1) that's a bit more typing, and (2) not everyone will remember to
do this.

Luckily, making everything pub(crate) is a pretty easy find-and-replace,
and can be quickly reverted once the dead code is deleted.

So this change is from doing a `s/pub /pub(crate) /g` (sans the few
symbols that really are public for api-test) and then deleting or
marking as cfg(test) all the dead code that remained, then reverting
the pub(crate) change. If code was totally unused, it was deleted. If it
was only used by tests, it was marked as cfg(test).

See merge request nvmetal/carbide!3592