chore: bump libredfish to 0.28.9.
We are temporarily disabling DPU FW updates in preingestion. This change in libredfish that we are pulling in will not return an error in machine_setup for BF3s that do not support enabling/disabling host rshim through redfish. This feature is not supported on older BMC FWs. We didnt want to put this logic in carbide because disabling the host rshim is nested in machine_setup. Carbide should know if any of the other calls fail.

See merge request nvmetal/carbide!3593