Fix N+1 queries for ManagedHostStateSnapshot
db::managed_host::load_by_machine_ids was issuing an additional SQL
query for every DPU id that was passed to it. This changes the way we
load managed host state snapshots to use a single SQL query, taking
advantage of postgres's JSON_AGG function in the same manner we did for
machines, machine_interfaces, etc.

To do this, we need to have a good way of converting postgres's JSON
representation into a Machine and an Instance. So this change makes it
so Machine and Instance are each deserialized from the database using
serde and sqlx::types::Json, using new types called
MachineSnapshotPgJson and InstanceSnapshotPgJson. These types' job is to
match the column names and types from postgres exactly, with no business
logic and derived Serialize and Deserialize impls. The FromRow
implementation for Machine and Instance becomes a simple matter of
reading the (single) column to these JSON types.

This makes it so a ManagedHostStateSnapshot query can return a JSON
object for the host machine and an array of JSON objects for each DPU,
and we just use plain old serde deserialization to read them into rust.
To prevent duplicate code, the FromRow path for all other Machine and
Instance queries are also retrofitted to use `row_to_json`, so that
reading via JSON is always the way we get this information from
postgres.

See merge request nvmetal/carbide!3598