feat: do not preingest DPUs if they are at the BMC & CEC versions corresponding to DOCA 2.5
If the CEC version is lower than `00.02.0182.0000_n02`, it requires a host power cycle to activate the new CEC. So, do not try preingesting DPUs if the BMC fw is new enough to support the redfish functionality we expect. The BMC fw associated with DOCA 2.5 (`BF-23.10-5`) satisfies all base redfish functionality we expect.

See merge request nvmetal/carbide!3608