feat(api): Make VPC isolation behavior configurable
This adds a new config file option `vpc_isolation_behavior` which can be used to customize VPC isolation within a site.

It defaults to `mutual_isolation` (the only behavior prior to this) but can be set to `open`, for sites where that's needed (pending the VPC peering feature becoming available).

The current implementation of the `open` behavior just skips including the `site_fabric_prefixes` members in the copy of `deny_prefixes` that gets sent to the DPU agents.

See merge request nvmetal/carbide!3625