Don't do so much cloning of ForgeClientConfig
It's a common pattern for API clients to build a single
ForgeClientConfig that lasts a long time (coming from env vars or a
config file), and then make several connections to carbide through the
lifetime of the process. But they have to clone the ForgeClientConfig
every time to do this. Fix this by allowing passing a ref.

To do this, make ForgeTlsClient and ApiConfig hold a borrowed ref to
ForgeClientConfig, rather than an owned instance.

See merge request nvmetal/carbide!3645