fix: Add OpenTelemetry DPU agent to renew mTLS certs
Retrying [this reverted MR](https://gitlab-master.nvidia.com/nvmetal/carbide/-/merge_requests/3470) with the same purpose: It adds `forge-dpu-otel-agent` to update client mTLS certs for the OpenTelemetry collector on the DPU. The new agent copies certs from `/opt/forge` (the ones used by `forge-dpu-agent`) and renews them on the same schedule as `forge-dpu-agent` and with the same expiry. The idea is that cert renewal of the OpenTelemetry collector will not depend on a running `forge-dpu-agent`.

This fixes the previous attempt by copying the certs in the otel agent rather than in postinst. If the `forge-dpu-agent` certs don't exist yet, the agent polls until they show up before copying. If the certs are present but not modified recently, the agent will copy them anyway (in case the otel agent hasn't run in a while).

The other side of this change ([here](https://gitlab-master.nvidia.com/nvmetal/forged/-/merge_requests/2929)) adds client authentication to the otlp endpoint on the site controller.

See merge request nvmetal/carbide!3602