MLT: Fix host creds
**Machine Lifecycle Test**

Since Forge no longer changes the host BMC username to 'root', we need to account for Lenovos which default to 'USERID'.

The creds are stored in corp vault as 'root:\<password\>' so this code change simply ignores the 'root' part and retrieves the password. The username is determined based on the machine vendor type.

See merge request nvmetal/carbide!3657