Machine-a-tron: Mock the server's desired firmware version
Instead of having to go and rewrite nvidia_dpu.tar.gz every time we bump
the firmware version (!3516, !3581), make the reported firmware versions
configurable, and if not configured, ask the server what firmware
versions it wants, and then mocks those versions in response.

Making a new API call to get desired firmware versions seems to be the
best option: If it were just the integration tests, we don't perform
firmware updates anyway, and even if we did we could easily just query
the database (or some other backchannel thing) to configure the desired
versions... but we also want to solve manually running machine-a-tron
against an existing cluster (including the in-kubernetes version of
machine-a-tron that runs when configured.) For that, we need it to relay
the right firmware versions so that ingestion works correctly.

For testing, enable firmware updates in machine-a-tron integration
tests (they fail if the reported versions are too old.)

See merge request nvmetal/carbide!3653