ci: disable in-pipeline scans
Move back to scheduled scans based on a known tag.  The in-pipeline
scans will generate a new scan configuration for every single container.
So when a CVE inevitably shows up, we get notifications for _all_ past
containers.  This just keeps scanning the latest container, but we don't
call it latest, we call it 'to-be-scanned' so you can't inadvertently
pull it.

See merge request nvmetal/carbide!3700