chore: MLT - Change maintenance check to health alert check
**Machine lifecycle test**

Due to this change: https://gitlab-master.nvidia.com/nvmetal/carbide/-/merge_requests/3670 we have to change how we check for machine in updating state.

Instead of checking for maintenance mode, now we must check for the absence of a specific health alert 'HostUpdateInProgress' to indicate that a machine is not undergoing a DPU FW update and is therefore safe to test.

See merge request nvmetal/carbide!3711