chore: Remove Managed-Host Details page on Admin Web UI
The managed host details page duplicated a lot of information with the Machine details pages. However it was never as complete as any of these, and less structured. Therefore operators often clicked the links on the page one more time to get to the Machine Details page.

This change removes the Managed Host Details page, and adjusts any links to it to the Host Machine Details page.

See merge request nvmetal/carbide!3714