chore: TDD baseline to capture metrics behavior before updating otel
There are a *ton* of changes in !3688, and we don't have nearly enough
test coverage to ensure that all the metrics we write out are going to
be unchanged.

This adds a bunch of new assertions to existing tests, and adds new
tests, to compare the result of doing metrics collection with known
baselines. I tried to make sure every place we collect metrics is
covered here.

The baseline were captured by writing out the exported prometheus text
via std::fs::write in the unit test, then changing the test to assert
that the metrics match.

To fix cases when metrics are "noisy" and don't reliably contain the
same values (like in latency metrics), this adds a simple parser,
ParsedPrometheusMetrics, which has a custom PartialEq implementation
which will ignore things we know are going to change between runs.

- For the various state controllers where we already have good test
  coverage, this captures a baseline at the end of each test.
- For other places, create new tests which build metrics structs from
  simple values, and capture baselines from there. In cases like these
  we just need to do a simple string-compare against the baseline, since
  there's no noise in the synthetic data.

With this, rebasing !3688 against this baseline will ensure that zero of
our prometheus gauges have any issues. (Full disclosure, I found exactly
2 bugs in !3688 using this, and exactly 1 bug in trunk because of this.)

See merge request nvmetal/carbide!3716