chore: Update opentelemetry to 0.28
We were stuck on opentelemetry 0.24 due to a [bug][0] in 0.25 and later,
which was fixed in 0.28. The bug is actually covered by our own test in
api/src/logging/setup.rs, so it's pretty simple to verify it's fixed...
once we upgrade the entire codebase.

Upgrading this one was a doozie.

OpenTelemetry no longer lets you add a callback to an entire meter. You
have to add the callback at the place you're defining the observable
instrument itself.

This requires a lot of refactoring... before, we would have
`emit_gauges()` methods which would read the metrics struct all at once
and then call `observer.observe()` for each metric in the struct,
typically doing clever things like pushing/popping from a vec of
attributes as we go. We can't do that any more.

Now, we have to register a callback for each gauge individually, and
each of these gauges has to get the current metrics object from a Mutex,
get the metric it wants, and observe just that one metric. So, we have
to unroll a lot of the stuff we were doing emit_gauges() into individual
gauge observers.

But we also have to incorporate the logic of using a `hold_until`
Duration and not emitting metrics if they're too stale, as well as logic
to incorporate a "fresh" attribute if the metrics are younger than a
certain age. So, this adds SharedMetricsHolder, which is a wrapper
around a Metrics struct that (1) keeps it in a mutex, (2) attaches a
date to it, (3) conditionally allows callers to see the metrics
object or not based on whether hold_until has expired, and (4) adds a
"fresh" attribute if configured to and the freshness period has not
expired.

Also, the observable gauges do not need to be stored anywhere. Just
building them is fine. Since we don't have emit_gauges any more, we can
drop the `instruments()` method too, and we don't have to store the
gauges in the metrics holders. Some of them have now become empty
structs. Some have dropped the struct altogether and just have a
`fn hydrate_meter()` method.

[0]: https://github.com/open-telemetry/opentelemetry-rust/issues/2213

See merge request nvmetal/carbide!3688