feat: Report disconnected IB devices in Machine Capabilities
When InfiniBand ports of a Host are not connected to the fabric, tenants can not use them during Instance creation. When a tenant tries to configure a port that is not connected, they will encounter the error message `UFM detected inactive state for GUID`.

However up to now, tenants had no way to know which ports are actually connected.

With this change, the Machine Capabilities properties of Machines will include a new `inactive_devices` list which will indicate which interfaces can't be used. The indices match the indices that a tenant would use in the `device_instance` field of the instance infiniband configuration while cerating the instance. The Forge UI could grey out changes that are using `inactive_devices`.

Some notes:
- `inactive_devices` has been used over `active_devices` in order to be backwards compatible. When the list is absent we should assume that the site has not been upgraded yet, and therefore potentially all devices are active.
- The approach of adding an `inactive_devices` field instead of e.g. just reporting the number of connected devices in the capability list has also been chosen for backward compatibility: Tenants would not need to adapt the device interface indices (`device_instance`) that they send during instance creation compared to before.

A host with 2 IB NICs where each of the NICs has the first port connected will be signaled with the following capability after this change:
```json
{
    vendor: "Mellanox Technologies",
    name: "MT2910 Family [ConnectX-7]",
    count: 4,
    inactive_devices: [0, 2]
}
```

See merge request nvmetal/carbide!3720