fixes for `sku show` and `sku assign`
This fixes the sku id type used for `sku show` . it was originally a uuid, but is now a string. this was missed in that change.

Also allows a sku id to be assigned to a machine when there isn't one and the machine is in ready state. originally this wasn't possible, but with the `ignore_unassigned_machines` config option, this is now possible.

See merge request nvmetal/carbide!3729