feat: Validate InfiniBand devices during SKU validation
With this change, the InfiniBand devices on a discovered Machine will be compared to the expected set of InfiniBand devices that is defined on the SKU.

For each vendor and device name combination of InfiniBand devices, the amount of expected InfiniBand devices is checked against the actually discovered InfiniBand devices.

In addition to just checking the presence of the devices, the connection status of InfiniBand devices is also verified. A Host with 2 IB NICs and 2 ports each where of of each ports is disconnected will not match against a SKU where all 4 ports are expected to be connected.

Also fixes a bug where the list of `inactive_devices` in Machine Capabilities was incorrectly generated due to not sorting the IB devices by PCI Slot.

See merge request nvmetal/carbide!3732