chore: Make test_integration use MachineATron
test_integration was testing very old code paths, and it was getting
difficult to support mocks which don't report the latest firmware/etc.

The main reason for not just deleting the test, is that it was asserting
on metrics lines in a way the current machine-a-tron test was not. So
this migrates those assertions to use machine-a-tron.

The other thing the old test was doing, is simulating a DPU agent
upgrade, which machine-a-tron doesn't do. Support for this is tracked in
FORGE-5738.

See merge request nvmetal/carbide!3776