feat: Store Health History for Machines
In order to allow operators to assess the state of certain Machines over time, it is useful to retain a history of health states.

With this change, the state handler will record the current health state of a Machine on every iteration (every 30s) into a new table `machine_health_history`.

If the health state did not change since the last iteration, no record is made.

A new API `FindMachineHealthHistories` can be used to query the health state of Machines over time. The new API has been chosen over adding a `health_history` field to `message Machine` in order to allow checking the health state also for deleted Machines.

A proof of concept visualization for Health History has been added to carbide-web. This will be extended in the future.

See merge request nvmetal/carbide!3773