fix(api,agent): Separate site fabric prefixes from deny_prefixes
This MR...
- Stops site_fabric_prefixes from being combined with deny_prefixes.
- Renames the original deny_prefixes field and continues to populate the old field with the combined set for older agents.
- Updates the protos to add site_fabric_prefixes as a separate field in the network config that the DPU agent receives.
- Updates the protos to send along VPC isolation behavior type to the DPU as well.
- Updates the FNN, pre-FNN etv, and pre-NVUE etv templates to add in site_fabric_prefixes to the deny policies when necessary.
  - If a security group is applied, site_fabric_prefixes is ignored. (default deny is applied if no security group rule matches.)

Where this gets a little weird is at the point where we lie to ourselves about supporting multiple tenants on a single machine.  To keep changes minimal, I have the SiteFabricPrefixes used in the tenant loop.  That loop is only ever going to have one iteration as things are now, so this should be fine, but we should consider reorganizing the FNN template, while we have the chance, to make it clearer how the policies lists are evaluated (e.g., they actually go first by the alphanumerically sorted order of the names and then by the actual policy priority.)

See merge request nvmetal/carbide!3768