feat: Support reusing single long-lived gRPC client connection
This makes a new wrapper for ForgeClientT which supports every method on
the forge.Forge API, but with a few benefits:

- Supports lazy connection to the server (connecting on first use, not
  on ::new)
- Doesn't require wrapping everything in tonic::Request, and calls
  .to_inner() for you (something that almost exactly 100% of clients
  were doing themselves... except for admin-cli's ping function which
  introspected the gRPC status from the result.)
- Is Send+Sync, so no need to be passed around as &mut (uses a tokio
  Mutex underneath)

The last point is a big one, it lets us hold an instance of ApiClient
in long-lived places, since it doesn't need to be `&mut`.

A big risk here is that we don't really know how good tonic+hyper is at
reliably reconnecting a long-lived re-used connection. So this just
migrates Machine-a-tron to using the new client type for now, until we
get a chance to learn more about the behavior.

See merge request nvmetal/carbide!3660