fix(api,agent): Provide an option to enable/disable stateful ACL support for NSGs
This MR..
- Plumbs in an option to let us enable/disable stateful ACL support. It defaults to true.

We spent a bunch of time with the HBN team testing stateful ACLs, and things should "just work," but we should have at least a basic lever, at least initially, to shut off the support if we need to.

A separate MR actually adds the stateful support in config.  This MR just provides a way for us to decide whether to even apply it.

See merge request nvmetal/carbide!3800