Use telemetry/pause provided by the BFB
During some investigation on the default BFB, it was noticed that
there is a local copy of both pause and telemetry as well as a script
to import them and start telemetry. So we now use this instead of
relying on the dpu build which adds external resources.

HBN removed it from their default yaml files provided in their bundle,
which is why this change came to be.

See merge request nvmetal/carbide!3807