fix: useless histogram forge_hardware_health_iteration_latency
Currently the `forge_hardware_health_iteration_latency` histogram is a useless metric because the only populated bucket is `le="+Inf"`. The next bucket is `le="10000"`, which is how often it takes 10 seconds or less to scrape all machines. Even after changing `forge-hw-health` to scrape machines concurrently, that never happens.

This MR fixes the histogram to use buckets for scrape intervals that happen in real life.

See merge request nvmetal/carbide!3803