Generate convenience From<> impls for RPC types
This puts the codegen work on ForgeApiClient into its own crate, and adds
some quality-of-life features for converting to Message types with only
a single field, or zero fields.

The separate crate is not specific to forge and could be open-sourced
(or upstreamed into tonic-build) if we want to.

Also, where possible in the admin and machine-a-tron crates, delete any
hand-written wrapper methods that are no longer serving any purpose,
letting callers just use ForgeApiClient directly.

See merge request nvmetal/carbide!3806