fix: improved differentiation of connection errors during exploration
https://jirasw.nvidia.com/browse/FORGE-5773

`site-explorer` scans IPMITOR network ranges, attempting to scrape the
Redfish API of each IP it finds. As it collects data, it of course will
pair hosts and DPUs together by way of matching serial numbers exposed
by the host's hardware inventory and the DPU BMC. However, before these
pairings can be made, it doesn't know what is what, and will also do
things like scan the DPU admin interface, which does not expose a Redfish
API.

In today's world, this means that unpaired/unmanaged/undiscovered machines
will report the DPU admin interface as an "unreachable" BMC, when it is
in fact not a BMC at all.

The idea here is to generalize some of that, by being more transparent
about the type of underlying issue that `site-explorer` is seeing
when attempting to scrape an IP address, e.g.
- `ConnectionRefused`: It has connectivity to the IP address, but the
connection was refused. A good indicator it doesn't have a Redfish API,
but can also mean there was a TLS handshake error.
- `ConnectionTimeout`: It has no connectivity to the endpoint IP, OR
port 443 is hanging out twiddling its thumbs.
- `Unreachable`: This remains as a catchall for other `NetworkError`
issues that came up as part of `reqwest` failing to get back some type
of HTTP response from the endpoint.

See merge request nvmetal/carbide!3814